requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
pillow==10.1.0
protobuf==4.25.1
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
httpx>=0.24.0
pyYaml==6.0
google-cloud-documentai 
google-cloud-storage